refactor(hotkey-router): split monolithic router into concern-scoped files
The hotkey_router.cc monolith (~1300 lines) is hard to navigate and edit
safely. Split it into smaller files, each owning one concern, so future
changes touch a narrower surface.

Pure code motion. No behavior change. Public API in hotkey_router.h is
byte-identical. Build verified via `ax cycle releasedbg`:
hookCount=20 installed=19 skipped=1 (baseline).

New files (all under mods/src/patches/):
  hotkey_router_action_table.h         constexpr action arrays
  hotkey_router_modifier_query.{h,cc}  Win32 modifier / KeyCode probes
  hotkey_router_dispatch_cache.{h,cc}  per-frame dispatch plan cache
  hotkey_router_trace_log.{h,cc}       trace-log gates and formatters
  hotkey_router_runtime_query.{h,cc}   runtime-binding lookup helpers
  hotkey_router_native_fleet_guard.cc  native-fleet guard state + delegates

Existing files:
  hotkey_router.h                      unchanged
  hotkey_router_native_fleet_guard.h   appended new free-fn declarations
  hotkey_router.cc                     slimmed to frame loop + entry hooks

mods/xmake.lua already globs src/**.cc, so no build-system change.
